refactor: use shared isLegacyTizen from platform
